change semantics of return value of Insert()
Breaking API change.  However, I'm likely the only user of this library.

Insert now returns true if the element was previously preset in the filter.
The count of items in the bloom filter is incremented only if at least
one of the bits was 0 (i.e., the element was *not* previously in the
filter).